+++ lib/input/toggle.d.ts
@@ -6,8 +6,7 @@ type TCheckboxWidget = IPropWidget<ICheckboxField>;
 export declare class ToggleInput extends BaseWidget<TCheckboxWidget> {
     protected readonly onIcon: keyof TSubset<IConfig, TIcon>;
     protected readonly offIcon: keyof TSubset<IConfig, TIcon>;
-    private toggleWrapper;
-    private toggleInner;
+    private toggle;
     view({ attrs: { field, value: val } }: CVnode<TCheckboxWidget>): m.Vnode<any, any> | m.Vnode<import("../interface/widget").IPropLayoutWidget<import("../interface/widget").IField>, unknown>;
 }
 export {};